Add web API routes for write/read log messages to monitor
* Implement write/read log API routes to monitor
* Fix connection client
* Start implementation of write mon log for syscheck/health
* Attach new frontend

+++ caen_tools/connection/client.py
@@ -84,7 +84,7 @@ class AsyncClient(BaseClient):
         connect_address = self.connect_addresses[receipt.executor]
 
         if receive_time is not None:
-            s.setsockopt(zmq.RCVTIMEO, receive_time * 1000)
+            s.setsockopt(zmq.RCVTIMEO, int(receive_time * 1000))
 
         with s.connect(connect_address) as sock:
             await sock.send_multipart([b"", receipt_str])
